* adding dev branch

* adding dev branch

* Make sure no caching happens for endpoints

* Allow quantity when charging

* (linked #2) add test frameworks

* (linked #2) add unit tests for utils

* (linked #2) configure npm commands to run mocha

* (linked #2) configure app to be able to run mocha tests

* (closes #2) add integration tests for posts

* add babel cli

* change api url

* Add procfile for heroku

* Adding travis yml file

* Bump bcrypt from 4.0.1 to 5.0.0

Bumps [bcrypt](https://github.com/kelektiv/node.bcrypt.js) from 4.0.1 to 5.0.0.
- [Release notes](https://github.com/kelektiv/node.bcrypt.js/releases)
- [Changelog](https://github.com/kelektiv/node.bcrypt.js/blob/master/CHANGELOG.md)
- [Commits](https://github.com/kelektiv/node.bcrypt.js/compare/v4.0.1...v5.0.0)


* Bump lodash from 4.17.15 to 4.17.20

Bumps [lodash](https://github.com/lodash/lodash) from 4.17.15 to 4.17.20.
- [Release notes](https://github.com/lodash/lodash/releases)
- [Commits](https://github.com/lodash/lodash/compare/4.17.15...4.17.20)


* remove travis

* fix procfile for heroku

* update gitignore

* added db import/export help in package json

* (closes #14, closes #4) add email templates and implement email service (#15)

* Change unsubscribe link

* Bump bl from 2.2.0 to 2.2.1 (#16)

Bumps [bl](https://github.com/rvagg/bl) from 2.2.0 to 2.2.1.
- [Release notes](https://github.com/rvagg/bl/releases)
- [Commits](https://github.com/rvagg/bl/compare/v2.2.0...v2.2.1)



* Comment unused user api endpoints (#17)

+++ src/routes/user.ts
@@ -1,33 +1,32 @@
 import express, { Request, Response } from 'express';
-// import request from 'request';
-import { DB } from '../database/types';
+import { DB, IDeleteResponse } from '../database/types';
 import UserService from '../services/user.service';
 import { buildApiPrefix, didDelete } from '../utils';
-import environment from '../utils/environment';
+// import environment from '../utils/environment';
 import passport from 'passport';
-import Stripe from 'stripe';
-import { IUserDocument, Query } from '../database/users/users.types';
+// import Stripe from 'stripe';
+import { Query } from '../database/users/users.types';
 import LogService from '../services/log.service';
 import EmailService, { IEmail, EmailType } from '../services/email.service';
-import { IReceiptDocument, IReceipt } from '../database/receipts/receipts.types';
-import ReceiptService from '../services/receipt.service';
+// import { IReceiptDocument, IReceipt } from '../database/receipts/receipts.types';
+// import ReceiptService from '../services/receipt.service';
 
-const CHARGE = 400;
+// const CHARGE = 400;
 
-const stripe = new Stripe(environment.STRIPE_KEY, {
-    apiVersion: '2020-03-02',
-});
+// const stripe = new Stripe(environment.STRIPE_KEY, {
+//     apiVersion: '2020-03-02',
+// });
 
 let userService: UserService;
 let logService: LogService;
 let emailService: EmailService;
-let receiptService: ReceiptService;
+// let receiptService: ReceiptService;
 
 export const user = (app: express.Application, db: DB): void => {
     userService = new UserService(db);
     logService = new LogService(db);
     emailService = new EmailService();
-    receiptService = new ReceiptService(db);
+    // receiptService = new ReceiptService(db);
 
     // USERS GET ROUTES
     app.get(
@@ -36,17 +35,17 @@ export const user = (app: express.Application, db: DB): void => {
         checkTokenHandler,
     );
 
-    app.get(buildApiPrefix('user', ''), passport.authenticate('userJWT', { session: false }), getUserHandler);
+    // app.get(buildApiPrefix('user', ''), passport.authenticate('userJWT', { session: false }), getUserHandler);
 
     app.get(buildApiPrefix('users', ''), passport.authenticate('adminJWT', { session: false }), getUsersHandler);
 
     // USERS POST ROUTES
-    app.post(buildApiPrefix('user', '/login'), loginHandler);
+    // app.post(buildApiPrefix('user', '/login'), loginHandler);
 
-    app.post(buildApiPrefix('user', '/charge'), chargeHandler);
+    // app.post(buildApiPrefix('user', '/charge'), chargeHandler);
 
     // client register
-    app.post(buildApiPrefix('user', '/register'), registerHandler);
+    // app.post(buildApiPrefix('user', '/register'), registerHandler);
 
     app.post(buildApiPrefix('user', '/subscribe'), subscribeHandler);
 
@@ -54,21 +53,21 @@ export const user = (app: express.Application, db: DB): void => {
     app.post(buildApiPrefix('user', ''), passport.authenticate('adminJWT', { session: false }), saveHandler);
 
     // validate password
-    app.post(
-        buildApiPrefix('user', '/validate'),
-        passport.authenticate('userJWT', { session: false }),
-        validatePasswordHandler,
-    );
+    // app.post(
+    //     buildApiPrefix('user', '/validate'),
+    //     passport.authenticate('userJWT', { session: false }),
+    //     validatePasswordHandler,
+    // );
 
-    app.post(buildApiPrefix('user', '/receipt'), receiptHandler);
+    // app.post(buildApiPrefix('user', '/receipt'), receiptHandler);
 
     app.post(buildApiPrefix('user', '/unsubscribe'), unsubscribeHandler);
 
     // USERS PATCH ROUTES
-    app.patch(buildApiPrefix('user', ''), passport.authenticate('userJWT', { session: false }), updateHandler);
+    // app.patch(buildApiPrefix('user', ''), passport.authenticate('userJWT', { session: false }), updateHandler);
 
     // client delete account
-    app.delete(buildApiPrefix('user', '/client'), passport.authenticate('userJWT', { session: false }), deleteHandler);
+    // app.delete(buildApiPrefix('user', '/client'), passport.authenticate('userJWT', { session: false }), deleteHandler);
 
     // admin delete user account
     app.delete(buildApiPrefix('user', '/admin'), passport.authenticate('adminJWT', { session: false }), deleteHandler);
@@ -78,19 +77,19 @@ const checkTokenHandler = (req: Request, res: Response) => {
     return res.json({ success: true, user: req.user });
 };
 
-const getUserHandler = async (req: Request, res: Response) => {
-    try {
-        const reqUser = req.user as IUserDocument;
-        const user = await userService.findOneWithQuery({ _id: reqUser._id }, false);
-        if (user) {
-            return res.json({ success: true, user });
-        }
-        return res.json({ success: false });
-    } catch (err) {
-        logService.createLog(err.stack, err.message, 'GET /api/user', err.code, 'ERROR');
-        return res.json({ success: false });
-    }
-};
+// const getUserHandler = async (req: Request, res: Response) => {
+//     try {
+//         const reqUser = req.user as IUserDocument;
+//         const user = await userService.findOneWithQuery({ _id: reqUser._id }, false);
+//         if (user) {
+//             return res.json({ success: true, user });
+//         }
+//         return res.json({ success: false });
+//     } catch (err) {
+//         logService.createLog(err.stack, err.message, 'GET /api/user', err.code, 'ERROR');
+//         return res.json({ success: false });
+//     }
+// };
 
 const getUsersHandler = async (_: Request, res: Response) => {
     try {
@@ -102,68 +101,65 @@ const getUsersHandler = async (_: Request, res: Response) => {
     }
 };
 
-const loginHandler = async (req: Request, res: Response) => {
-    // log user in
-    try {
-        const { email, password } = req.body;
-        const user = await userService.authorize(email, password);
-        if (!user) {
-            res.status(401);
-            res.send('Unauthorized request!');
-            return;
-        }
-        const result = userService.login(user);
-        return res.json({ success: true, access_token: result.access_token, user: result.user });
-    } catch (err) {
-        logService.createLog(err.stack, err.message, 'POST /api/user/login', err.code, 'ERROR');
-        return res.json({ success: false, message: `Error authorizing user!` });
-    }
-};
+// const loginHandler = async (req: Request, res: Response) => {
+//     // log user in
+//     try {
+//         const { email, password } = req.body;
+//         const user = await userService.authorize(email, password);
+//         if (!user) {
+//             res.status(401);
+//             res.send('Unauthorized request!');
+//             return;
+//         }
+//         const result = userService.login(user);
+//         return res.json({ success: true, access_token: result.access_token, user: result.user });
+//     } catch (err) {
+//         logService.createLog(err.stack, err.message, 'POST /api/user/login', err.code, 'ERROR');
+//         return res.json({ success: false, message: `Error authorizing user!` });
+//     }
+// };
 
-const registerHandler = async (req: Request, res: Response) => {
-    // register user
-    try {
-        req.body.role = 'user';
-        const user = await userService.save(req.body);
-        if (user) {
-            const email: IEmail = {
-                body: 'Some body', // TODO
-                title: 'Some title', // TODO
-                from: environment.FROM_EMAIL,
-                to: user.email,
-                type: EmailType.UserRegistration,
-            };
-            emailService.sendEmail(email);
+// const registerHandler = async (req: Request, res: Response) => {
+//     // register user
+//     try {
+//         req.body.role = 'user';
+//         const user = await userService.save(req.body);
+//         if (user) {
+//             const email: IEmail = {
+//                 subject: 'Account Registration',
+//                 to: user.email,
+//                 type: EmailType.UserRegistration,
+//             };
 
-            const result = userService.login(user);
-            return res.json({ success: true, access_token: result.access_token, user: result.user });
-        }
-        logService.createLog('N/A', 'User was null when registering', 'POST /api/user/register', 'N/A', 'WARN');
-        return res.json({ success: false, message: `Error creating user!` });
-    } catch (err) {
-        logService.createLog(err.stack, err.message, 'POST /api/user/register', err.code, 'ERROR');
-        return res.json({ success: false, message: `Error creating user!` });
-    }
-};
+//             emailService.sendEmail(email);
+
+//             const result = userService.login(user);
+//             return res.json({ success: true, access_token: result.access_token, user: result.user });
+//         }
+//         logService.createLog('N/A', 'User was null when registering', 'POST /api/user/register', 'N/A', 'WARN');
+//         return res.json({ success: false, message: `Error creating user!` });
+//     } catch (err) {
+//         logService.createLog(err.stack, err.message, 'POST /api/user/register', err.code, 'ERROR');
+//         return res.json({ success: false, message: `Error creating user!` });
+//     }
+// };
 
 const subscribeHandler = async (req: Request, res: Response) => {
     // subscribe user
     try {
-        // TODO send an email
         req.body.role = 'subscriber';
+        req.body.notify = true;
         const user = await userService.save(req.body);
         if (user) {
-            const email: IEmail = {
-                body: 'Some body', // TODO
-                title: 'Some title', // TODO
-                from: environment.FROM_EMAIL,
-                to: user.email,
-                type: EmailType.UserSubscription,
-            };
-            emailService.sendEmail(email);
+            // const email: IEmail = {
+            //     subject: 'New Subscriber',
+            //     to: user.email,
+            //     type: EmailType.UserSubscription,
+            // };
+            // emailService.sendEmail(email);
             return res.json({ success: true });
         }
-        logService.createLog('N/A', 'User was null when registering', 'POST /api/user/register', 'N/A', 'WARN');
+        logService.createLog('N/A', 'User was null when subscribing', 'POST /api/user/subscribe', 'N/A', 'WARN');
         return res.json({ success: false, message: `Error creating user!` });
     } catch (err) {
         logService.createLog(err.stack, err.message, 'POST /api/user/subscribe', err.code, 'ERROR');
@@ -182,87 +178,85 @@ const saveHandler = async (req: Request, res: Response) => {
     }
 };
 
-const updateHandler = async (req: Request, res: Response) => {
-    // update user
-    try {
-        const user = await userService.save(req.body);
+// const updateHandler = async (req: Request, res: Response) => {
+//     // update user
+//     try {
+//         const user = await userService.save(req.body);
 
-        if (user) {
-            const email: IEmail = {
-                body: 'Some body', // TODO
-                title: 'Some title', // TODO
-                from: environment.FROM_EMAIL,
-                to: user.email,
-                type: EmailType.UserAccountChange,
-            };
-            emailService.sendEmail(email);
-        }
+//         if (user) {
+//             const email: IEmail = {
+//                 subject: 'Account Change',
+//                 to: user.email,
+//                 type: EmailType.UserAccountChange,
+//             };
+//             emailService.sendEmail(email);
+//         }
 
-        return res.json({ success: true, user: user });
-    } catch (err) {
-        logService.createLog(err.stack, err.message, 'PATCH /api/user', err.code, 'ERROR');
-        return res.json({ success: false, message: `Error updating user!` });
-    }
-};
+//         return res.json({ success: true, user: user });
+//     } catch (err) {
+//         logService.createLog(err.stack, err.message, 'PATCH /api/user', err.code, 'ERROR');
+//         return res.json({ success: false, message: `Error updating user!` });
+//     }
+// };
 
-const validatePasswordHandler = async (req: Request, res: Response) => {
-    try {
-        const result = await userService.validatePassword(req.body.email, req.body.password);
-        return res.json({
-            success: result,
-        });
-    } catch (err) {
-        logService.createLog(err.stack, err.message, 'POST /api/user/validate', err.code, 'ERROR');
-        return res.json({ success: false, message: `Error validating password!` });
-    }
-};
+// const validatePasswordHandler = async (req: Request, res: Response) => {
+//     try {
+//         const result = await userService.validatePassword(req.body.email, req.body.password);
+//         return res.json({
+//             success: result,
+//         });
+//     } catch (err) {
+//         logService.createLog(err.stack, err.message, 'POST /api/user/validate', err.code, 'ERROR');
+//         return res.json({ success: false, message: `Error validating password!` });
+//     }
+// };
 
-const chargeHandler = async (req: Request, res: Response) => {
-    try {
-        const paymentIntent = await stripe.paymentIntents.create({
-            amount: CHARGE * req.body.quantity,
-            currency: 'usd',
-            // Verify your integration in this guide by including this parameter
-            metadata: { integration_check: 'accept_a_payment' },
-        });
-        return res.json({ success: true, paymentIntent });
-    } catch (err) {
-        logService.createLog(err.stack, err.message, 'POST /api/user/charge', err.code, 'ERROR');
-        return res.json({ success: false, message: `Error creating payment intent` });
-    }
-};
+// const chargeHandler = async (req: Request, res: Response) => {
+//     try {
+//         const paymentIntent = await stripe.paymentIntents.create({
+//             amount: CHARGE * req.body.quantity,
+//             currency: 'usd',
+//             // Verify your integration in this guide by including this parameter
+//             metadata: { integration_check: 'accept_a_payment' },
+//         });
+//         return res.json({ success: true, paymentIntent });
+//     } catch (err) {
+//         logService.createLog(err.stack, err.message, 'POST /api/user/charge', err.code, 'ERROR');
+//         return res.json({ success: false, message: `Error creating payment intent` });
+//     }
+// };
 
-const receiptHandler = async (req: Request, res: Response) => {
-    try {
-        const { receipt_email: email, amount, description } = req.body.data.object;
-        const isTest = description === '(created by Stripe CLI)';
-        const receipt: IReceipt = {
-            email,
-            amount,
-            item: 'donation',
-            deleted: false,
-        };
-        if (isTest || (email && amount)) {
-            // send email thanking user for donation
-            const newReceipt = await receiptService.save(<IReceiptDocument>receipt);
-            if (newReceipt) {
-                const emailToSend: IEmail = {
-                    body: 'Some body', // TODO
-                    title: 'Some title', // TODO
-                    from: environment.FROM_EMAIL,
-                    to: email || 'test@gmail.com',
-                    type: EmailType.UserDontationReceipt,
-                };
-                emailService.sendEmail(emailToSend);
-            }
-        }
+// const receiptHandler = async (req: Request, res: Response) => {
+//     try {
+//         const { receipt_email: email, amount, description } = req.body.data.object;
+//         const isTest = description === '(created by Stripe CLI)';
+//         const receipt: IReceipt = {
+//             email,
+//             amount,
+//             item: 'donation',
+//             deleted: false,
+//         };
+//         if (isTest || (email && amount)) {
+//             // send email thanking user for donation
+//             const newReceipt = await receiptService.save(<IReceiptDocument>receipt);
+//             if (newReceipt) {
+//                 if (!!email) {
+//                     const emailToSend: IEmail = {
+//                         subject: 'Donation Receipt',
+//                         to: email,
+//                         type: EmailType.UserDontationReceipt,
+//                     };
+//                     emailService.sendEmail(emailToSend);
+//                 }
+//             }
+//         }
 
-        return res.json({ success: true });
-    } catch (err) {
-        logService.createLog(err.stack, err.message, 'POST /api/user/receipt', err.code, 'ERROR');
-        return res.json({ success: false, message: `Error creating receipt` });
-    }
-};
+//         return res.json({ success: true });
+//     } catch (err) {
+//         logService.createLog(err.stack, err.message, 'POST /api/user/receipt', err.code, 'ERROR');
+//         return res.json({ success: false, message: `Error creating receipt` });
+//     }
+// };
 
 const deleteHandler = async (req: Request, res: Response) => {
     try {
@@ -271,9 +265,7 @@ const deleteHandler = async (req: Request, res: Response) => {
             const deleteResult = await userService.deleteOne(req.query);
             if (didDelete(deleteResult) && user) {
                 const email: IEmail = {
-                    body: 'Some body', // TODO
-                    title: 'Some title', // TODO
-                    from: environment.FROM_EMAIL,
+                    subject: 'Account Deletion',
                     to: user.email,
                     type: EmailType.UserAccountDeletion,
                 };
@@ -291,13 +283,30 @@ const deleteHandler = async (req: Request, res: Response) => {
 const unsubscribeHandler = async (req: Request, res: Response) => {
     try {
         if (req.body.email) {
-            const query: Query = { email: req.body.email, role: 'subscriber' };
-            const deleteResult = await userService.deleteOne(query);
+            const query: Query = { email: req.body.email };
+            let user = await userService.findOneWithQuery(query, false);
+            let deleteResult: IDeleteResponse = {};
+            if (user) {
+                if (user.role == 'subscriber') {
+                    query.role = 'subscriber';
+                    deleteResult = await userService.deleteOne(query);
+                } else if (user.role == 'user') {
+                    user.notify = false;
+                    user = await userService.save(user);
+                    if (user) {
+                        // const email: IEmail = {
+                        //     subject: 'Account Change',
+                        //     to: user.email,
+                        //     type: EmailType.UserAccountChange,
+                        // };
+                        // emailService.sendEmail(email);
+                        return res.json({ success: true, message: `Unsubscribed user successfully.` });
+                    }
+                }
+            }
             if (didDelete(deleteResult)) {
                 const email: IEmail = {
-                    body: 'Some body', // TODO
-                    title: 'Some title', // TODO
-                    from: environment.FROM_EMAIL,
+                    subject: 'Unsubscribe',
                     to: req.body.email,
                     type: EmailType.UserAccountDeletion,
                 };
